Update pmap.py
- Added menu system with "screens"

+++ pmap.py
@@ -15,17 +15,22 @@ from PIL import ImageDraw   # type: ignore
 from PIL import ImageFont   # type: ignore
 import ST7789               # type: ignore
 
-
 # imports used for buttons + temperature sensing
 from gpiozero import Button # type: ignore
 from gpiozero import CPUTemperature # type: ignore
 from gpiozero import PWMLED # type: ignore
 
+# import used to send terminal commands
 import os
 
-# global vars
+# Global Variables
 
-airplay_status_text = ""
+airplay_status = 1 # on
+screen = 1 # home screen
+font = ImageFont.truetype("Ubuntu-Regular.ttf", 30)
+font_small = ImageFont.truetype("Ubuntu-Regular.ttf", 20)
+icons = ImageFont.truetype("pmap_icons.ttf", 30)
+
 
 # initialize INA219
 ina219 = INA219(addr=0x43)
@@ -61,27 +66,63 @@ y_but = Button(24)
 
 
 def a_pressed():
-    os.system('sudo nqptp &')
-    os.system('sudo shairport-sync &')
-    global airplay_status_text
-    airplay_status_text = "AirPlay Enabled"
+    global screen
+
+    if screen == 1:
+        screen = 5
+    elif screen > 1 and screen < 6:
+        screen = 1
+    elif screen == 5:
+        screen = 1
 
 def b_pressed():
-    os.system('sudo pkill nqptp')
-    os.system('sudo pkill shairport-sync')
-    global airplay_status_text
-    airplay_status_text = "AirPlay Disabled"
+
+    global airplay_status
+    global screen
+
+    if screen == 1 and airplay_status == 0:
+        os.system('sudo nqptp &')
+        os.system('sudo shairport-sync &')
+        airplay_status = 1
+    elif screen == 1 and airplay_status == 1:
+        os.system('sudo pkill nqptp')
+        os.system('sudo pkill shairport-sync')
+        airplay_status = 0
+
+    if screen == 3:
+        screen = 6
+        time.sleep(1)
+        os.system('sudo shutdown now')
 
 def x_pressed():
-    os.system('sudo reboot now')
+    global screen
+    if screen == 1:
+        screen = 2
+
+    if screen == 3:
+        os.system('sudo reboot now')
 
 def y_pressed():
-    os.system('sudo shutdown now')
+    global screen
+    if screen == 1:
+        screen = 3
 
 a_but.when_pressed = a_pressed
 b_but.when_pressed = b_pressed
 x_but.when_pressed = x_pressed
 y_but.when_pressed = y_pressed
+
+# Icons
+icon_settings = "\uE801"
+icon_batt_100 = "\uF240"
+icon_batt_75 = "\uF241"
+icon_batt_50 = "\uF242"
+icon_batt_25 = "\uF243"
+icon_batt_0 = "\uF244"
+icon_airplay = "\uE814"
+icon_power = "\uE810"
+icon_left_arrow = "\uE806"
+icon_reload = "\uE809"
 
 
 def draw_rotated_text(image, text, position, angle, font, fill=(255, 255, 255)):
@@ -99,13 +140,9 @@ def draw_rotated_text(image, text, position, angle, font, fill=(255, 255, 255)):
     # Paste the text into the image, using it as a mask for transparency.
     image.paste(rotated, position, rotated)
 
-
-
-while True:
-
+def battery_stats():
     # Getting Battery Values
     bus_voltage = ina219.getBusVoltage_V()             # voltage on V- (load side)
-    shunt_voltage = ina219.getShuntVoltage_mV() / 1000 # voltage between V+ and V- across the shunt
     current = ina219.getCurrent_mA()                   # current in mA
     power = ina219.getPower_W()                        # power in W
     p = (bus_voltage - 3)/1.2*100
@@ -116,43 +153,200 @@ while True:
     #print("PSU Voltage:   {:6.3f} V".format(bus_voltage + shunt_voltage))
     #print("Shunt Voltage: {:9.6f} V".format(shunt_voltage))
 
-    lv = "Load Voltage:  {:6.3f} V".format(bus_voltage)
-    curr = "Current:       {:6.3f} A".format(current/1000)
-    pow = "Power:         {:6.3f} W".format(power)
-    perc = "Percent:       {:3.1f}%".format(p)
+    lv = round(bus_voltage,3)
+    curr = round(current/1000,3)
+    pow = round(power,3)
+    perc = round(p,0)
 
-    # Print values to terminal
-    print(lv)
-    print(curr)
-    print(pow)
-    print(perc)
-    print("")
+    return lv,curr,pow,perc
 
+def cpu_temp():
     cpu = CPUTemperature()
 
-    cpu_temp = "CPU Temp:     "+str(cpu.temperature)
+    return round(cpu.temperature,2)
 
+def render_screen_1(): # Home Screen
+    
+    #Getting battery info
+    batt = battery_stats()
+    curr = batt[1]    
+    perc = batt[3]
+    
     # Clear the display to a black background.
     img = Image.new('RGB', (WIDTH, HEIGHT), color=(0, 0, 0))
     draw = ImageDraw.Draw(img)
 
-    # Load Fonts
-    font = ImageFont.truetype("Ubuntu-Regular.ttf", 15)
-    icons = ImageFont.truetype("pmap_icons.ttf", 30)
+    # Refer to global font variables
+    global font 
+    global icons 
 
-    draw_rotated_text(img, lv, (0, 0), 0, font, fill=(255, 255, 255))
-    draw_rotated_text(img, curr, (0, 20), 0, font, fill=(255, 255, 255))
-    draw_rotated_text(img, pow, (0, 40), 0, font, fill=(255, 255, 255))
-    draw_rotated_text(img, perc, (0, 60), 0, font, fill=(255, 255, 255))
-    draw_rotated_text(img, cpu_temp, (0, 100), 0, font, fill=(255, 255, 255))
-    draw_rotated_text(img, airplay_status_text, (0, 160), 0, font, fill=(255, 255, 255))
+    # Top Left - Settings Icon
+    draw_rotated_text(img, icon_settings, (0, 0), 0, icons, fill=(255, 255, 255))
 
-    draw_rotated_text(img, "\uF240 \uF1BC \uE814 \uF2C7", (0, 120), 0, icons, fill=(255, 255, 255))
+    # Bottom Left - AirPlay Icon
+    if airplay_status == 1:
+        draw_rotated_text(img, icon_airplay, (0, 200), 0, icons, fill=(127, 255, 127))
+    elif airplay_status == 0:
+        draw_rotated_text(img, icon_airplay, (0, 200), 0, icons, fill=(255, 127, 127))
+    else:
+        draw_rotated_text(img, icon_airplay, (0, 200), 0, icons, fill=(255, 255, 255))
+
+    # Top Right - Battery Icon
+    fill_col = (255, 255, 255) #white, default
+    if curr > 0: #charging
+        fill_col = (127, 255, 127) #green, charging
+    elif perc < 10: #not charging, batt < 10%
+        fill_col = (255, 127, 127) #red, needs to charge
+  
+    if perc > 90:
+        draw_rotated_text(img, icon_batt_100, (200, 0), 0, icons, fill=fill_col)
+    elif perc > 75:
+        draw_rotated_text(img, icon_batt_75, (200, 0), 0, icons, fill=fill_col)
+    elif perc > 50:
+        draw_rotated_text(img, icon_batt_50, (200, 0), 0, icons, fill=fill_col)
+    elif perc > 25:
+        draw_rotated_text(img, icon_batt_25, (200, 0), 0, icons, fill=fill_col)
+    else:
+        draw_rotated_text(img, icon_batt_0, (200, 0), 0, icons, fill=fill_col)
+    
+    # Bottom Right - Power Icon
+    draw_rotated_text(img, icon_power, (200, 200), 0, icons, fill=(255, 255, 255))
+    
+    
+    draw_rotated_text(img, "pmap", (120, 120), 0, font, fill=(255, 255, 255))
+
+    # Write buffer to display hardware, must be called to make things visible on the
+    # display!
+    disp.display(img)    
+
+def render_screen_2(): # Battery Screen
+    
+    batt = battery_stats()
+    
+    lv = batt[0]
+    curr = batt[1]    
+    pow = batt[2]
+    perc = batt[3]
+
+    lv_text = "Load Voltage:  {:6.3f} V".format(lv)
+    curr_text = "Current:       {:6.3f} A".format(curr)
+    pow_text = "Power:         {:6.3f} W".format(pow)
+    perc_text = "Percent:       {:3.1f}%".format(perc)
+    temp_text = "CPU Temp:     "+str(cpu_temp())
+    
+    # Print values to terminal
+    print(lv_text)
+    print(curr_text)
+    print(pow_text)
+    print(perc_text)
+    print(temp_text)
+    print("")
+    
+    # Clear the display to a black background.
+    img = Image.new('RGB', (WIDTH, HEIGHT), color=(0, 0, 0))
+    draw = ImageDraw.Draw(img)
+
+    # Refer to global font variables
+    global font 
+    global icons 
+
+    # Top left - Back Icon
+    draw_rotated_text(img, icon_left_arrow, (0, 0), 0, icons, fill=(255, 255, 255))
+
+    draw_rotated_text(img, "Battery", (40, 0), 0, font, fill=(255, 255, 255))
+
+    draw_rotated_text(img, lv_text, (10, 50), 0, font_small, fill=(255, 255, 255))
+    draw_rotated_text(img, curr_text, (10, 70), 0, font_small, fill=(255, 255, 255))
+    draw_rotated_text(img, pow_text, (10, 90), 0, font_small, fill=(255, 255, 255))
+    draw_rotated_text(img, perc_text, (10, 110), 0, font_small, fill=(255, 255, 255))
+
+    # Write buffer to display hardware, must be called to make things visible on the
+    # display!
+    disp.display(img)
+
+def render_screen_3(): # Power Screen
+    
+    # Clear the display to a black background.
+    img = Image.new('RGB', (WIDTH, HEIGHT), color=(0, 0, 0))
+    draw = ImageDraw.Draw(img)
+
+    # Refer to global font variables
+    global font 
+    global icons 
+
+    # Top left - Back Icon
+    draw_rotated_text(img, icon_left_arrow, (0, 0), 0, icons, fill=(255, 255, 255))
+
+    draw_rotated_text(img, "Power", (40, 0), 0, font, fill=(255, 255, 255))
+
+    # Bottom Left - Shutdown Icon
+    draw_rotated_text(img, icon_power, (0, 200), 0, icons, fill=(255, 127, 127))
+
+    # Top Right - Reboot Icon
+    fill_col = (255, 255, 255) #white, default
+    draw_rotated_text(img, icon_reload, (200, 0), 0, icons, fill=fill_col)
 
 
     # Write buffer to display hardware, must be called to make things visible on the
     # display!
     disp.display(img)
 
+def render_screen_5(): # Settings Screen
+    
+    # Clear the display to a black background.
+    img = Image.new('RGB', (WIDTH, HEIGHT), color=(0, 0, 0))
+    draw = ImageDraw.Draw(img)
 
-    time.sleep(2)
+    # Refer to global font variables
+    global font 
+    global icons 
+
+    # Top left - Back Icon
+    draw_rotated_text(img, icon_left_arrow, (0, 0), 0, icons, fill=(255, 255, 255))
+
+    draw_rotated_text(img, "Settings", (40, 0), 0, font, fill=(255, 255, 255))
+
+    draw_rotated_text(img, "No Settings yet :)", (10, 50), 0, font_small, fill=(255, 255, 255))
+
+    # Write buffer to display hardware, must be called to make things visible on the
+    # display!
+    disp.display(img)
+
+def render_screen_6(): # Shutdown Complete Screen
+    
+    # Clear the display to a black background.
+    img = Image.new('RGB', (WIDTH, HEIGHT), color=(0, 0, 0))
+    draw = ImageDraw.Draw(img)
+
+    # Refer to global font variables
+    global font 
+    global icons 
+
+    draw_rotated_text(img, "Shutting down", (40, 0), 0, font, fill=(255, 255, 255))
+
+    draw_rotated_text(img, "Wait for RPi status light", (10, 50), 0, font_small, fill=(255, 255, 255))
+    draw_rotated_text(img, "to turn off before", (10, 70), 0, font_small, fill=(255, 255, 255))
+    draw_rotated_text(img, "switching off power", (10, 90), 0, font_small, fill=(255, 255, 255))
+
+    # Write buffer to display hardware, must be called to make things visible on the
+    # display!
+    disp.display(img)
+
+
+while True:
+
+    match screen: # type: ignore
+        case 1:
+            render_screen_1()
+        case 2:
+            render_screen_2()
+        case 3:
+            render_screen_3()
+        case 5:
+            render_screen_5()
+        case 6:
+            render_screen_6()
+        case _:
+            ...
+
+    time.sleep(1)
